+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Change Log
 * Realiza ajustes no acoplamento para funcionar apenas com MUDOU_POSICAO (PL-3355)
 * Refatora updates de km dos processos para usarem function centralizadora (PL-3290)
 * Adiciona erro especifico ao tentar por hubodômetro em veículos motorizados (PL-3386)
+* Adiciona validação para impedir que veículos acoplados sejam inativados (PL-3397)
 
 ### Bug Fixes
 * Corrige diferenças entre branches na edição de veículo (PL-3326)